(BSR)[PRO] Ends nicos work and adapt with searchIndividualOffer
