+++ pro/cypress/e2e/features/collectiveBooking.feature
@@ -5,36 +5,46 @@ Feature: Search for collective bookings
     Given I am logged in
     And I go to "Réservations" page
     And I go to "Réservations collectives" page
-    And I display offers
-    When I search for Offre with name "offer 39"
-    Then I should see "1" results
-    And I should see my Offre  
+    When I display offers
+    And I search for "Offre" with text "offer 39"
+    Then These results should be displayed
+      | Réservation | Nom de l'offre | Établissement                               | Places et prix | Statut  |
+      |          80 | offer 39       | ECOLE ELEMENTAIRE PUBLIQUE FRANCOIS MOISSON |  25 places100€ | annulée |
 
   Scenario: It should find collective bookings by establishments
     Given I am logged in
     And I go to "Réservations" page
     And I go to "Réservations collectives" page
-    And I display offers
-    When I search for Établissement with name "LYCEE POLYVALENT METIER ROBERT DOISNEAU"
-    Then I should see "1" results
-    And I should see my Établissement
+    When I display offers
+    And I search for "Établissement" with text "LYCEE POLYVALENT METIER ROBERT DOISNEAU"
+    Then These results should be displayed
+      | Réservation | Nom de l'offre | Établissement                           | Places et prix | Statut      |
+      |          79 | offer 38       | LYCEE POLYVALENT METIER ROBERT DOISNEAU |  25 places100€ | annulée     |
+      |          72 | offer 31       | LYCEE POLYVALENT METIER ROBERT DOISNEAU |  25 places100€ | annulée     |
+      |          65 | offer 24       | LYCEE POLYVALENT METIER ROBERT DOISNEAU |  25 places100€ | remboursée  |
+      |          58 | offer 17       | LYCEE POLYVALENT METIER ROBERT DOISNEAU |  25 places100€ | terminée    |
+      |          51 | offer 10       | LYCEE POLYVALENT METIER ROBERT DOISNEAU |  25 places100€ | confirmée   |
+      |          44 | offer 3        | LYCEE POLYVALENT METIER ROBERT DOISNEAU |  25 places100€ | préréservée |
 
   Scenario: It should find collective bookings by booking number
     Given I am logged in
     And I go to "Réservations" page
     And I go to "Réservations collectives" page
-    And I display offers
-    When I search for Numéro de réservation with number "66"
-    Then I should see "1" results
-    And I should see my Établissement    
+    When I display offers
+    And I search for "Numéro de réservation" with text "66"
+    Then These results should be displayed
+      | Réservation | Nom de l'offre | Établissement                               | Places et prix | Statut  |
+      |          66 | offer 25       | ECOLE ELEMENTAIRE PUBLIQUE FRANCOIS MOISSON |  25 places100€ | annulée |
 
-  # pas fini
   Scenario: It should find collective bookings with two filters
     Given I am logged in
     And I go to "Réservations" page
     And I go to "Réservations collectives" page
-    And I fill Lieu with ""
+    When I fill venue with "real_venue 1 eac_2_lieu [BON EAC]"
     And I display offers
-    When I search for Numéro de réservation with number "66"
-    Then I should see "1" results
-    And I should see my Établissement      
+    And I search for "Établissement" with text "ECOLE ELEMENTAIRE PUBLIQUE FRANCOIS MOISSON"
+    Then These results should be displayed
+      | Réservation | Nom de l'offre | Établissement                               | Places et prix | Statut    |
+      |          80 | offer 39       | ECOLE ELEMENTAIRE PUBLIQUE FRANCOIS MOISSON |  25 places100€ | annulée   |
+      |          66 | offer 25       | ECOLE ELEMENTAIRE PUBLIQUE FRANCOIS MOISSON |  25 places100€ | annulée   |
+      |          52 | offer 11       | ECOLE ELEMENTAIRE PUBLIQUE FRANCOIS MOISSON |  25 places100€ | confirmée |
