+++ pro/src/screens/Bookings/BookingsRecapTable/Filters/FilterByOmniSearch.tsx
@@ -111,6 +111,7 @@ export const FilterByOmniSearch = ({
         name="omnisearch-criteria"
         className={styles['omnisearch-filter-select']}
         disabled={isDisabled}
+        data-testid="select-omnisearch-criteria"
         onBlur={handleOmniSearchCriteriaChange}
         onChange={handleOmniSearchCriteriaChange}
         value={selectedOmniSearchCriteria}
@@ -128,6 +129,7 @@ export const FilterByOmniSearch = ({
         className={styles['omnisearch-filter-input']}
         disabled={isDisabled}
         id="text-filter-input"
+        data-testid="omnisearch-filter-input-text"
         onChange={handleOmniSearchChange}
         placeholder={placeholderText}
         value={keywords}
